Refactor ChatInterface and Sidebar integration in home route
